Changed the way ThreadRipper.java functions
Uhh so yeah now it operates like an object because it was too bothersome
to try and initialize things that I needed to keep adding to add
features. Now it has everything initialize nicely and the filePath is
decided upon creation of the object. Also its a lot smarter now. If it
has seen a duplicate image with a name it will not download it again.